Feat: 팀참가시 참가 팀 목록 api 연결
