+++ types/agenda/team/teamDataTypes.ts
@@ -4,6 +4,8 @@ export interface teamDataProps {
   teamName: string;
   teamLeaderIntraId: string;
   teamMateCount: number;
-  teamKey: string;
+  teamKey?: string;
+  teamAward?: string;
+  awardPriority?: number;
   coalitions: string[] | Coalition[];
 }
